YAY-21 refactor: move profile page to dynamic route
